add Button의 AddMode 구현중
